ml-service/requirements.txt
flask==2.3.3
flask-cors==4.0.0
opencv-python==4.8.1.78
ultralytics==8.0.186
easyocr==1.7.0
numpy==1.24.3
pillow==10.0.0
python-dotenv==1.0.0
requests==2.31.0